:fire: remove hello world
